Changed from a play bin to pipeline + added volume control
Swapped from using an all inclusive playbin to using a custom pipeline
with a filesrc, decoder, etc. This will facilitate using a RTP stream as
audio output instead of a audio sink.

Volume control was also added as an element in the pipeline. The decoder
writes to the volume control to allow the server to control volume of
clients before RTP streaming.

Signed-off-by: Alex Waterman <imNotListening@gmail.com>