Have prototype RTP/RTPC streaming for the stream server.
This server will require a client written using GST sadly, since there
is no session management. The client must be aware of what port it is
recv'ing data on and read the UDP RTP data from said port. This is
annoying, however, the client really does not need to be very
sophisticated.

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>