RTP/RTCP streaming verified to work
The addresses used for my laptop are hard coded into the code at the
moment so that will need to change. The very prelimary basics of the
server streaming API are also working. More will need to be added like
what to do on song end, etc, related to an actual music player.

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>