Adds first support for streaming API
Based on the API drawn on the white board this API should be capable of
streaming an RTP/RTCP session to a specific IP address. Check the
stream_server.c file for an example implementation.

At the moment I have not put full support for music seeking/timing in.
Once I have a more clear idea of what exactly is needed regarding that
I will put in a function call back for music finishing, a timer, or both.

Signed-off-by: Alex Waterman <imNotListening@gmail.com>