Adds more support for streaming API
Added complete functions for getting the length and position of the music
stream. This is done though asynchronous calls to the pipeline. It is MP
safe according to the docs.

Also added some more macros to the music.h header in the hopes of making
monitoring the state of the pipeline easier. The macros are:

	MUSIC_IS_NOGO(state)
	MUSIC_IS_READY(state)
	MUSIC_IS_PUASED(state)
	MUSIC_IS_PLAYING(state)

They return true when the state that is passed is equivalent to the tested
state. E.g MUSIC_IS_PLAYING(state) evaluates to true when the pipeline is
actually streaming music. The macros call a function that will block if
the pipeline is in the process of transitioning state.

Signed-off-by: Alex Waterman <imNotListening@gmail.com>