Added pipe list structure for C, and started work on better encoding.
Since C does not have a nice STL, I have implemented a small list
structure to make ym life easier. This also is a part of the multi
streaming test code.

I have also started working on using a better encoding for the RTP
stream. As of now I have not figured out why it doesn't work but I
think it has something to do with setting the caps correctly in the
client. It could also be possible to use mplayer as the client if I
can figure out how to do rtsp.

Signed-off-by: Alex Waterman <imnotListening@gmail.com>