Moved the RTP streaming to plain UDP streaming
What was once done with RTP streamsing with a uLaw basec codec has now
been moved to using just plain UDP with a FLAC codec. It solves many an
issue with RTP, because RTP is a nightmare.

The only change as far as the API is concerned is the function

    int music_make_rtp_pipeline()

has been renamed to

    int music_make_pipeline()

And no longer accepts the second integer argument (rtcp_port) since it is
no longer relevant. The first argument, the rtp port is now just treated
as the UDP port and should not need any modifcation.

Signed-off-by: Alex Waterman <imNotListening@gmail.com>