Added a shell script wrapper to run programs
The shell script wrapper runs programs in the music project that link
against the music shared library. Its usage is simple:

    music_run.sh <command> [args]

This will set up the library load path, and pass any arguments specified
to the command <command>.

Also added a simple seek command to the stream_server test code. That
will soon be implemented as part of the music API to hide the details
from any calling code.

Also, also, added a simple hack in the top level make file to wipe the
extraneous files in bin/ when make clean is run. I may remove this
later, not sure yet.

Signed-off-by: Alex Waterman <imNotListening@gmail.com>