Wow, what a redisgn. Shifted the music lib to the client.
Instead of having all of the logic in the server, the logis has been
moved to the clients. The clients pull a stream from a web server
and play the stream. The client has access to typical music playing
functionality, e.g, play/pause, stop, new song, seek, etc. Only now
music quality is perfect, and the response times to seeking and
volume change is instantaneous. Pretty nifty.

Signed-off-by: Alex Waterman <imNotListening@gmail.com>